feat(cli): add cli login
- `@screenpipe/dev login` now lets you choose apiKey or cli login
